==> Notebook/1_ETL_de_dataset_para_FastAPI.ipynb <==
<h1>ETL de Datos de Películas: limpieza, normalización y estructuración de datos.</h1>

<h3>Librerías</h3>

#Importamos las librerías que vamos a usar
import pandas as pd
from pandas import json_normalize
import ast
from datetime import datetime

<h3>Carga de Datos</h3>

#definio la ruta del CVS
ruta_movies_dataset = r'C:\Users\Practical Tecno\Desktop\HENRY\datasets_org\movies_dataset.csv'
ruta_credits = r'C:\Users\Practical Tecno\Desktop\HENRY\datasets_org\credits.csv'

#Cargo el archivo CSV
df_movies = pd.read_csv(ruta_movies_dataset)
df_credit = pd.read_csv(ruta_credits)

<h3>Revisión de Datos</h3>

df_movies.head()

df_credit.head()

df_movies.info()

df_credit.info()

<h3>Limpieza de Datos</h3>

Eliminación de datos duplicados

df_movies = df_movies.drop_duplicates()
df_credit_cast = df_credit.drop_duplicates()

Eliminación de columnas

# Lista de columnas a eliminar
columnas_a_eliminar = ['adult', 'homepage','imdb_id', 'original_language', 'original_title', 'poster_path', 'video', 'tagline']

# Eliminar las columnas del DataFrame
df_movies= df_movies.drop(columns=columnas_a_eliminar)

#Los valores nulos del campo release date deben eliminarse.
df_movies.dropna(subset=['release_date'], inplace=True)

#Los datos que estén ausentes en las columnas serán eliminados si no representan más del 3% del total de registros. 
#Esto se realizará para evitar la eliminación de filas en su totalidad utilizando el método .dropna(), lo cual podría afectar significativamente el dataset.

df_movies.dropna(subset=['overview'], inplace=True)
df_movies.dropna(subset=['popularity'], inplace=True)
df_movies.dropna(subset=['production_companies'], inplace=True)
df_movies.dropna(subset=['production_countries'], inplace=True)
df_movies.dropna(subset=['revenue'], inplace=True)
df_movies.dropna(subset=['runtime'], inplace=True)
df_movies.dropna(subset=['spoken_languages'], inplace=True)
df_movies.dropna(subset=['status'], inplace=True)
df_movies.dropna(subset=['title'], inplace=True)
df_movies.dropna(subset=['vote_average'], inplace=True)
df_movies.dropna(subset=['vote_count'], inplace=True)

#En la columna de "status", se seleccionarán únicamente los registros que contengan el estado "Released". Estos registros son los que se consideran publicados en la plataforma 
#y serán utilizados en el sistema de recomendación. 
conteo_rumored = df_movies['status'].value_counts()['Rumored']
print(f"El número de registros con 'Rumored' en la columna 'status' es: {conteo_rumored}")

conteo_Post_Production = df_movies['status'].value_counts()['Post Production']
print(f"El número de registros con 'Post Production' en la columna 'status' es: {conteo_Post_Production}")

conteo_In_Production = df_movies['status'].value_counts()['In Production']
print(f"El número de registros con 'In Production' en la columna 'status' es: {conteo_In_Production}")

conteo_Canceled = df_movies['status'].value_counts()['Canceled']
print(f"El número de registros con 'Canceled' en la columna 'status' es: {conteo_Canceled}")


df_movies = df_movies.loc[~df_movies['status'].isin(['Rumored', 'Post Production', 'In Production', 'Canceled', 'Planned'])]

#Eliminar la columna 'status'
df_movies = df_movies.drop(columns='status')

<h2>Transformación de Datos</h2>

#Los valores nulos de los campos revenue, budget deben ser rellenados por el número 0.
df_movies['revenue'] = df_movies['revenue'].fillna(0)
df_movies['budget'] = df_movies['budget'].fillna(0)

#Transformamos los valores a int o float según corresponda.
df_movies['id'] = df_movies['id'].astype('int64')
df_movies['budget']=df_movies['budget'].astype('float')
df_movies['vote_count']=df_movies['vote_count'].astype('int')
df_movies['revenue']=df_movies['revenue'].astype('int')
df_movies['budget']=df_movies['budget'].astype('int')


#Las fechas deben tener el formato AAAA-mm-dd. Además, se creará la columna "release_year" donde se extraerá el año de la fecha de estreno, si está disponible.
df_movies['release_date'] = pd.to_datetime(df_movies['release_date'], errors='coerce').dt.strftime('%Y-%m-%d')

# Crear la columna release_year extrayendo el año de la fecha de estreno
df_movies['release_year'] = pd.to_datetime(df_movies['release_date'], errors='coerce').dt.year


# Para calcular el retorno de inversión (ROI), es necesario convertir los campos relevantes a tipo float.
def calculate_roi(row):
    try:
        revenue = (row['revenue'])
        budget = (row['budget'])
        return revenue / budget if budget != 0 else 0
    except (TypeError, ValueError):
        return 0

# Aplicar la función para calcular el ROI y crear la columna 'return'
df_movies['return'] = df_movies.apply(calculate_roi, axis=1)

<h3>Desanidado</h3>

#Para facilitar mi trabajo, he decidido desanidar las columnas, ya que creo que será más conveniente trabajar de esta manera. 
# Además, estoy creando una función que me permita expandir columnas en un DataFrame que contiene listas dentro de diccionarios.
def expand_columns(df, columns):
    for column in columns:
        df[column] = df[column].apply(lambda x: ast.literal_eval(x) if pd.notnull(x) else []) #Convierte el contenido de cada celda en la columna column de df de una cadena de texto que representa una lista de diccionarios a una lista de diccionarios Python, utilizando ast.literal_eval. Si el valor es nulo (pd.notnull(x)), se devuelve una lista vacía.
        # Expande la columna column en filas, de modo que cada elemento de la lista de diccionarios se convierte en una fila separada en df
        df = df.explode(column)
        # Normaliza la columna column después de expandirla, lo que significa que convierte cada diccionario dentro de la lista en una serie de columnas.
        col_df = pd.json_normalize(df[column])
        # Renombra las columnas normalizadas agregando el nombre original de la columna como prefijo, para evitar conflictos si hay nombres de columnas duplicados en diferentes listas de diccionarios.
        col_df = col_df.add_prefix(f'{column}_')
        # Elimina la columna original column del DataFrame df y la reemplaza con las columnas normalizadas (col_df), asegurándose de que el índice del DataFrame se reinicie para evitar problemas con el índice de las filas.
        df = df.drop(columns=[column]).reset_index(drop=True).join(col_df)
    return df

columns_to_expand = ['genres', 'production_companies', 'production_countries', 'spoken_languages','belongs_to_collection']

# Expandir las columnas de listas de diccionarios
df_movies = expand_columns(df_movies, columns_to_expand)

df_movies

#Si desanido el conjunto de datos "credit" en la versión que nos entregaron, obtengo 9 millones de filas. 
# Por lo tanto, planeo dividir la tabla en dos, manteniendo el ID en ambas partes para poder trabajar de manera más eficiente y ocupando menos espacio.

# Crear df_credit_cast con las columnas "cast" y "id"
df_credit_cast = df_credit[['cast', 'id']].copy()

# Crear df_credit_crew con las columnas "crew" y "id"
df_credit_crew = df_credit[['crew', 'id']].copy()


def expand_columns(df, columns):
    for column in columns:
        df[column] = df[column].apply(lambda x: ast.literal_eval(x) if pd.notnull(x) else []) #Convierte el contenido de cada celda en la columna column de df de una cadena de texto que representa una lista de diccionarios a una lista de diccionarios Python, utilizando ast.literal_eval. Si el valor es nulo (pd.notnull(x)), se devuelve una lista vacía.
        # Expande la columna column en filas, de modo que cada elemento de la lista de diccionarios se convierte en una fila separada en df
        df = df.explode(column)
        # Normaliza la columna column después de expandirla, lo que significa que convierte cada diccionario dentro de la lista en una serie de columnas.
        col_df = pd.json_normalize(df[column])
        # Renombra las columnas normalizadas agregando el nombre original de la columna como prefijo, para evitar conflictos si hay nombres de columnas duplicados en diferentes listas de diccionarios.
        col_df = col_df.add_prefix(f'{column}_')
        # Elimina la columna original column del DataFrame df y la reemplaza con las columnas normalizadas (col_df), asegurándose de que el índice del DataFrame se reinicie para evitar problemas con el índice de las filas.
        df = df.drop(columns=[column]).reset_index(drop=True).join(col_df)
    return df

columns_to_expand = ['cast']

df_credit_cast = expand_columns(df_credit_cast, columns_to_expand)

def expand_columns(df, columns):
    for column in columns:
        df[column] = df[column].apply(lambda x: ast.literal_eval(x) if pd.notnull(x) else []) #Convierte el contenido de cada celda en la columna column de df de una cadena de texto que representa una lista de diccionarios a una lista de diccionarios Python, utilizando ast.literal_eval. Si el valor es nulo (pd.notnull(x)), se devuelve una lista vacía.
        # Expande la columna column en filas, de modo que cada elemento de la lista de diccionarios se convierte en una fila separada en df
        df = df.explode(column)
        # Normaliza la columna column después de expandirla, lo que significa que convierte cada diccionario dentro de la lista en una serie de columnas.
        col_df = pd.json_normalize(df[column])
        # Renombra las columnas normalizadas agregando el nombre original de la columna como prefijo, para evitar conflictos si hay nombres de columnas duplicados en diferentes listas de diccionarios.
        col_df = col_df.add_prefix(f'{column}_')
        # Elimina la columna original column del DataFrame df y la reemplaza con las columnas normalizadas (col_df), asegurándose de que el índice del DataFrame se reinicie para evitar problemas con el índice de las filas.
        df = df.drop(columns=[column]).reset_index(drop=True).join(col_df)
    return df

columns_to_expand = ['crew']

df_credit_crew = expand_columns(df_credit_crew, columns_to_expand)

df_credit_cast

df_credit_crew

df_movies

<h3>Eliminación de datos después del Desanidado</h3>

# Eliminar duplicados
df_movies = df_movies.drop_duplicates()
df_credit_cast = df_credit_cast.drop_duplicates()
df_credit_crew = df_credit_crew.drop_duplicates()

# Lista de columnas a eliminar
columnas_a_eliminar = ['genres_id','production_companies_id', 'production_countries_iso_3166_1', 'spoken_languages_iso_639_1']

# Eliminar las columnas del DataFrame
df_movies= df_movies.drop(columns=columnas_a_eliminar)

# Lista de columnas a eliminar
columnas_a_eliminar = ['cast_cast_id', 'cast_credit_id', 'cast_gender','cast_id', 'cast_profile_path',]

# Eliminar las columnas del DataFrame
df_credit_cast= df_credit_cast.drop(columns=columnas_a_eliminar)

#Estoy eliminando los actores que no considero como principales.
df_credit_cast = df_credit_cast[df_credit_cast['cast_order'] <= 3]

columnas_a_eliminar = ['crew_credit_id','crew_department', 'crew_gender','crew_id', 'crew_profile_path']

# Eliminar las columnas del DataFrame
df_credit_crew= df_credit_crew.drop(columns=columnas_a_eliminar)

#En la columna "crew_job", vamos a conservar únicamente los registros donde el cargo sea "Director".
df_credit_crew = df_credit_crew[df_credit_crew['crew_job'] == 'Director']

#Elimino la columa crew_job que ya no aporta más información.
df_credit_crew = df_credit_crew.drop(columns='crew_job')

df_movies.dropna(subset=['genres_name'], inplace=True)
df_movies.dropna(subset=['production_countries_name'], inplace=True)
df_movies.dropna(subset=['production_companies_name'], inplace=True)
df_movies.dropna(subset=['spoken_languages_name'], inplace=True)

#Eliminamos las películas que sean más viejas que 1970
df_movies = df_movies[df_movies['release_year'] >= 1970]

#Elimino las filas duplicadas teniendo en cuenta únicamente las columnas "title", "genres_name" y "production_companies_name". 
# Creo que estas columnas serán útiles para el aprendizaje automático, mientras que otras como el lenguaje no son tan relevantes en este contexto.
df_movies = df_movies.drop_duplicates(subset=['title', 'genres_name', 'production_companies_name'])


df_movies['budget'] = pd.to_numeric(df_movies['budget'], errors='coerce')
df_movies['popularity'] = pd.to_numeric(df_movies['popularity'], errors='coerce')
df_movies['revenue'] = pd.to_numeric(df_movies['revenue'], errors='coerce')
df_movies['runtime'] = pd.to_numeric(df_movies['runtime'], errors='coerce')
df_movies['vote_average'] = pd.to_numeric(df_movies['vote_average'], errors='coerce')
df_movies['vote_count'] = pd.to_numeric(df_movies['vote_count'], errors='coerce')
df_movies['release_year'] = pd.to_numeric(df_movies['release_year'], errors='coerce')
df_movies['return'] = pd.to_numeric(df_movies['return'], errors='coerce')

df_credit_cast['cast_order'] = pd.to_numeric(df_credit_cast['cast_order'], errors='coerce')

# Guardar los DataFrames en formato Parquet
df_movies.to_parquet('movies_limpio.parquet', index=False)
df_credit_cast.to_parquet('credit_cast_limpio.parquet', index=False)
df_credit_crew.to_parquet('credit_crew_limpio.parquet', index=False)
