==> Notebook/4_Modelo_de_recomendación.ipynb <==
<h3>Sistema de recomendación:</h3>

Una vez que toda la data es consumible por la API, está lista para consumir por los departamentos de Analytics y Machine Learning, y nuestro EDA nos permite entender bien los datos a los que tenemos acceso, es hora de entrenar nuestro modelo de machine learning para armar un sistema de recomendación de películas. El EDA debería incluir gráficas interesantes para extraer datos, como por ejemplo una nube de palabras con las palabras más frecuentes en los títulos de las películas. Éste consiste en recomendar películas a los usuarios basándose en películas similares, por lo que se debe encontrar la similitud de puntuación entre esa película y el resto de películas, se ordenarán según el score de similaridad y devolverá una lista de Python con 5 valores, cada uno siendo el string del nombre de las películas con mayor puntaje, en orden descendente. Debe ser deployado como una función adicional de la API anterior y debe llamarse:

<b>def recomendacion(<i>titulo</i>)</b>: Se ingresa el nombre de una película y te recomienda las similares en una lista de 5 valores.

<h3>Librerías</h3>

import pandas as pd
import nltk
from nltk.corpus import stopwords
import string
from collections import Counter
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.feature_extraction.text import CountVectorizer


<h3>Carga de Datos</h3>

ruta_movies_dataset = r'C:\Users\Practical Tecno\Desktop\HENRY\Henry_Proyecto_Individual_1_Memolli\datasets\movies_limpio.parquet'
credit_crew_dataset = r'C:\Users\Practical Tecno\Desktop\HENRY\Henry_Proyecto_Individual_1_Memolli\datasets\credit_crew_limpio.parquet'
credit_cast_dataset = r'C:\Users\Practical Tecno\Desktop\HENRY\Henry_Proyecto_Individual_1_Memolli\datasets\credit_cast_limpio.parquet'

df_movies = pd.read_parquet(ruta_movies_dataset)
df_credit_crew = pd.read_parquet(credit_crew_dataset)
df_credit_cast = pd.read_parquet(credit_cast_dataset)

df_movies

df_movies = df_movies[df_movies['release_year'] >= 1980]

principales_idiomas = ['English', 'Français', 'Deutsch', 'Español', 'Italiano']

df_movies = df_movies[df_movies['spoken_languages_name'].isin(principales_idiomas)]

paises_principales = ['United States of America', 'United Kingdom', 'France', 'Canada', 
                      'Japan', 'Germany', 'Italy', 'Russia', 'India', 'Spain', 'Argentina']

df_movies = df_movies[df_movies['production_countries_name'].isin(paises_principales)]

<h3>Transformaciones previas a unir los Dataframes</h3>
• Los valores de las columnas "genres_name", "crew_name" y "cast_name" se convierten en listas, dejando solo un valor de ID en cada una de ellas.

genero = (df_movies[['id', 'genres_name']]
                               .drop_duplicates()
                               .groupby('id')['genres_name']
                               .apply(list)
                               .reset_index(name='generos'))

directores = (df_credit_crew[['id', 'crew_name']]
                               .drop_duplicates()
                               .groupby('id')['crew_name']
                               .apply(list)
                               .reset_index(name='directores'))

actores = (df_credit_cast[['id', 'cast_name']]
                               .drop_duplicates()
                               .groupby('id')['cast_name']
                               .apply(list)
                               .reset_index(name='actores'))

Se crea otro DataFrame que contiene selecciones de columnas de varios DataFrames diferentes.

col_movies = df_movies[['id', 'title','overview']]
col_genero = genero[['id', 'generos']]
col_directores = directores[['id', 'directores']]
col_actores = actores[['id', 'actores']]

# Realizar merge secuencial para combinar los DataFrames
df_tags = pd.merge(col_movies, col_genero, on='id', how='inner')
df_tags = pd.merge(df_tags, col_directores, on='id', how='inner')
df_tags = pd.merge(df_tags, col_actores, on='id', how='inner')

df_tags

df_tags.info()

<h3>Transformaciones</h3>
La columna 'overview' será transformada para dividir el texto en caracteres individuales.</br>
Eliminamos cualquier espacio que pueda existir. Este paso es crucial para la vectorización.</br>
Se crea la columna 'tags' para consolidar los resultados de las otras columnas, asegurándonos de tener solo uno por ID.

#Se Separas en caracteres a la columna de 'overview'
#Se eliminan las filas vacías que puedan haber quedado.. 
df_tags.dropna(subset=['overview'], inplace=True)
df_tags['overview'] = df_tags['overview'].apply(lambda x:x.split())

#Se ha creado una función para eliminar los espacios que puedan existir.
def eliminar_espacios(cadena):
    sin_espacios = []
    for i in cadena:
        sin_espacios.append(i.replace(" ", ""))
    return sin_espacios

#Se aplica la función creada para eliminar espacios en las columnas 'gener_name', 'crew_name' y 'cast_name'.
df_tags['generos'] = df_tags['generos'].apply(eliminar_espacios)
df_tags['directores'] = df_tags['directores'].apply(eliminar_espacios)
df_tags['actores'] = df_tags['actores'].apply(eliminar_espacios)


Se ha creado la columna "tags" para almacenar toda la información que fue transformada.

df_tags['etiquetas'] = df_tags['overview'] + df_tags['generos'] + df_tags['directores'] + df_tags['actores']

#Eliminamos las columnas que ya no nos sirven.
df_tags = df_tags.drop(columns = ['overview', 'generos', 'directores', 'actores','id'])

#la línea de código convierte cada lista de strings en la columna 'etiquetas' del DataFrame df_tags en una cadena única de texto, donde cada elemento está separado por espacios, 
# utilizando una función lambda con apply().
df_tags['etiquetas'] = df_tags ['etiquetas'].apply(lambda x: " ".join(x))

df_tags = df_tags.drop_duplicates()

df_tags

nltk.download('stopwords')
stop_words = set(stopwords.words('english'))


<h3>Creación de Modelo de Recomendación</h3>


#Reiniciar el index porque me estaban dando mal las recomendaciones. 
df_tags.reset_index(drop=True, inplace=True)

cv = CountVectorizer(max_features = 10000 , stop_words= 'english')

vector = cv.fit_transform(df_tags['etiquetas']).toarray()

vector.shape

similitud_coseno = cosine_similarity(vector)

similitud_coseno

def recomendacion(titulo):
    try:
        indice = df_tags[df_tags['title'] == titulo].index[0]
    
        distancia = sorted(list(enumerate(similitud_coseno[indice])), reverse = True, key = lambda x: x[1])

        recomendadas = [df_tags.iloc[i[0]].title for i in distancia[1:6]]
        
        print(f"Porque viste '{titulo}', tal vez te guste:")
        for pelicula in recomendadas:
            print(pelicula)
    except IndexError:
        print(f"No se encontró la película '{titulo}' en la base de datos.")

recomendacion("Toy Story")

#Descargamos la base de datos que utilizamos para crear la función de recomendación. 
df_tags.to_parquet('tags_ML.parquet', index=False)
